requirements.txt
absl-py==0.2.1
astor==0.6.2
bleach==1.5.0
#Django==2.0.5
gast==0.2.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.6
Markdown==2.6.11
numpy==1.14.3
opencv-python==3.4.1.15
Pillow==5.1.0
protobuf==3.5.2.post1
pymongo==3.6.1
pytz==2018.4
PyYAML==3.12
scipy==1.1.0
six==1.11.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
